+++ package.json
@@ -17,7 +17,7 @@
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
-    "test": "node node_modules/jest/bin/jest.js",
+    "test": "node node_modules/jest/bin/jest.js --verbose",
     "test-android": "cross-env NODE_ENV=test jest --config jest.android.json",
     "flow": "flow",
     "prettier": "prettier --write *.js *block-management/*.js",
